Refactor the file with helper functions
